compiler: f64 @data_sentinel (backport missed in the EMLX drop)
The no-observation sentinel was still :f32 while the default precision
moved to f64 when EMLX (the f32-only backend) was dropped. Match the dev
workspace: use a :f64 sentinel so no-data models don't inject an f32
scalar into the otherwise-f64 logp graph. compiler.ex is now in parity
with the dev workspace.

+++ lib/exmc/compiler.ex
@@ -9,9 +9,11 @@ defmodule Exmc.Compiler do
 
   alias Exmc.{IR, Rewrite, PointMap, Transform}
 
-  # Sentinel for models without observation data.
-  # A scalar constant — adds one trivial XLA parameter, compiles once.
-  @data_sentinel Nx.tensor(0.0, type: :f32, backend: Nx.BinaryBackend)
+  # Sentinel for models without observation data — a dummy scalar that is
+  # ignored for no-data models (the logp closure discards it). f64 to match the
+  # default precision after EMLX (the f32-only backend) was dropped; the runtime
+  # step/leapfrog paths use a precision-aware sentinel (see BatchedLeapfrog).
+  @data_sentinel Nx.tensor(0.0, type: :f64, backend: Nx.BinaryBackend)
 
   @doc """
   Compile an IR into a logp function and its PointMap.
